# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Clinic booking API (learning)
  description: |
    Hand-maintained OpenAPI spec. Update this file when routes change; UI at `/api/docs`.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local (default PORT from `.env`)
...
paths:
...
  /api/v1/appointments:
    get:
      tags: [Appointments]
      summary: List appointments for the authenticated patient (same data as `/my`)
      operationId: listAppointmentsByPatient
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List (may be empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not a patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
    post:
      tags: [Appointments]
      summary: Book a slot for the authenticated patient (transactional)
      operationId: bookAppointment
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookAppointmentRequest"
      responses:
        "201":
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not a patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Slot or patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: Slot already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/appointments/waitlist:
    post:
      tags: [Waitlist]
      summary: Join waitlist for a booked (unavailable) slot
      operationId: joinSlotWaitlist
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slotId]
              properties:
                slotId:
                  type: integer
                  minimum: 1
      responses:
        "201":
          description: Waitlist row created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaitlistEntry"
        "400":
          description: Slot still available (`SLOT_STILL_AVAILABLE`) or validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not a patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Slot or patient not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "409":
          description: Already on waitlist (`WAITLIST_DUPLICATE`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/appointments/waitlist/me:
    get:
      tags: [Waitlist]
      summary: List caller waitlist rows with slot times
      operationId: listMyWaitlist
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List (may be empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WaitlistEntryWithSlot"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not a patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/appointments/waitlist/{waitlistId}:
    delete:
      tags: [Waitlist]
      summary: Remove own waitlist entry
      operationId: deleteWaitlistEntry
      security:
        - bearerAuth: []
      parameters:
        - name: waitlistId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  removed:
                    type: boolean
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not the owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Entry not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/appointments/my:
    get:
      tags: [Appointments]
      summary: List appointments for the authenticated patient
      operationId: listMyAppointments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Caller appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not a patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/appointments/doctor:
    get:
      tags: [Appointments]
      summary: List appointments on slots owned by the authenticated doctor
      operationId: listDoctorAppointments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Appointments for doctor slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not a doctor or doctor not linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/appointments/{id}/cancel:
    patch:
      tags: [Appointments]
      summary: Cancel appointment (authenticated patient); frees the slot
      description: Caller must be the patient who owns the appointment. Allowed source statuses — `pending`, `confirmed`.
      operationId: cancelAppointment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Appointment updated to cancelled; slot available again
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not the owning patient
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "422":
          description: Invalid status transition (e.g. already cancelled)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/appointments/{id}/confirm:
    patch:
      tags: [Appointments]
      summary: Doctor confirms a pending appointment
      operationId: confirmAppointment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Appointment confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not slot owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "422":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/appointments/{id}/reject:
    patch:
      tags: [Appointments]
      summary: Doctor rejects a pending appointment (frees slot)
      operationId: rejectAppointment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Appointment rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not slot owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "422":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/appointments/{id}/cancel-as-doctor:
    patch:
      tags: [Appointments]
      summary: Doctor cancels a pending or confirmed visit (frees the slot)
      description: Caller must own the slot. Allowed source statuses — `pending`, `confirmed`; result status `cancelled`.
      operationId: cancelAppointmentAsDoctor
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Appointment cancelled; slot available again
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "403":
          description: Not slot owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "422":
          description: Invalid status transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /api/v1/appointments/{id}:
    get:
      tags: [Appointments]
      summary: Get appointment by id
      operationId: getAppointmentById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Appointment row
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
        "404":
          description: Appointment not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Use Authorization header — `Bearer <token>` from `/api/v1/auth/login`
...
  schemas:
...
    WaitlistEntry:
      type: object
      properties:
        id:
          type: integer
        slotId:
          type: integer
        patientId:
          type: integer
        createdAt:
          type: string
          format: date-time
    WaitlistEntryWithSlot:
      allOf:
        - $ref: "#/components/schemas/WaitlistEntry"
        - type: object
          properties:
            slotStartTime:
              type: string
              format: date-time
            slotEndTime:
              type: string
              format: date-time
            doctorId:
              type: integer
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        slotId:
          type: integer
        patientId:
          type: integer
        status:
          type: string
          example: pending
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        deletedAt:
          type: string
          nullable: true
    BookAppointmentRequest:
      type: object
      required: [slotId]
      properties:
        slotId:
          type: integer
          minimum: 1
          description: Visit is booked for the authenticated patient (`req.user.id`).
...
    ErrorBody:
      type: object
      properties:
        errorCode:
          type: string
        message:
          type: string
        requestId:
          type: string
          format: uuid